printresult double
printresult double

+++ real/src/main/java/com/your/dream/real/AppView.java
@@ -72,7 +72,7 @@ public class AppView {
 		return iReturn;
 	}
 	
-	public void printResult(int bill){
+	public void printResult(double bill){
 		//Print bill to User
 		System.out.println("Your total charges are $"+bill);
 	}
